requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
absl-py=2.3.1=pypi_0
addict=2.4.0=pypi_0
basicsr=1.0.0=dev_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
causal-conv1d=1.0.0=pypi_0
certifi=2022.12.7=pypi_0
charset-normalizer=2.1.1=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
einops=0.8.1=pypi_0
expat=2.7.1=h6a678d5_0
filelock=3.13.1=pypi_0
fonttools=4.59.0=pypi_0
fsspec=2024.6.1=pypi_0
future=1.0.0=pypi_0
grpcio=1.74.0=pypi_0
hf-xet=1.1.5=pypi_0
huggingface-hub=0.34.3=pypi_0
idna=3.4=pypi_0
imageio=2.37.0=pypi_0
jinja2=3.1.4=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
lmdb=1.7.3=pypi_0
lpips=0.1.4=pypi_0
mamba-ssm=1.0.1=pypi_0
markdown=3.8.2=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.5=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h7934f7d_0
networkx=3.3=pypi_0
ninja=1.11.1.4=pypi_0
numpy=1.24.3=pypi_0
opencv-python=4.12.0.88=pypi_0
openssl=3.0.17=h5eee18b_0
packaging=25.0=pypi_0
pillow=11.0.0=pypi_0
pip=25.1=pyhc872135_2
platformdirs=4.3.8=pypi_0
protobuf=6.31.1=pypi_0
ptflops=0.7.4=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pyparsing=3.2.3=pypi_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0.post0=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2025.7.33=pypi_0
requests=2.28.1=pypi_0
safetensors=0.5.3=pypi_0
scikit-image=0.25.2=pypi_0
scipy=1.15.3=pypi_0
setuptools=78.1.1=py310h06a4308_0
six=1.17.0=pypi_0
sqlite=3.50.2=hb25bd0a_1
sympy=1.13.3=pypi_0
tb-nightly=2.21.0a20250802=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tifffile=2025.5.10=pypi_0
timm=1.0.19=pypi_0
tk=8.6.14=h993c535_1
tokenizers=0.21.4=pypi_0
tomli=2.2.1=pypi_0
torch=2.1.0+cu118=pypi_0
torchaudio=2.1.0+cu118=pypi_0
torchvision=0.16.0+cu118=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.54.1=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025b=h04d1e81_0
urllib3=1.26.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yapf=0.43.0=pypi_0
zlib=1.2.13=h5eee18b_1
